+++ lib/segment/src/index/field_index/numeric_index/mod.rs
@@ -18,7 +18,7 @@ use rocksdb::DB;
 use serde_json::Value;
 
 use self::immutable_numeric_index::ImmutableNumericIndex;
-use self::numeric_index_key::{Encodable, NumericIndexKey};
+use self::numeric_index_key::NumericIndexKey;
 use super::utils::check_boundaries;
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::common::rocksdb_wrapper::DatabaseColumnWrapper;
@@ -28,11 +28,60 @@ use crate::index::field_index::stat_tools::estimate_multi_value_selection_cardin
 use crate::index::field_index::{
     CardinalityEstimation, PayloadBlockCondition, PayloadFieldIndex, PrimaryCondition, ValueIndexer,
 };
+use crate::index::key_encoding::{
+    decode_f64_key_ascending, decode_i64_key_ascending, encode_f64_key_ascending,
+    encode_i64_key_ascending,
+};
 use crate::telemetry::PayloadIndexTelemetry;
 use crate::types::{FieldCondition, FloatPayloadType, IntPayloadType, PayloadKeyType, Range};
 
 const HISTOGRAM_MAX_BUCKET_SIZE: usize = 10_000;
 const HISTOGRAM_PRECISION: f64 = 0.01;
+
+pub trait Encodable: Copy {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8>;
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self);
+
+    fn cmp_encoded(&self, other: &Self) -> std::cmp::Ordering;
+}
+
+impl Encodable for IntPayloadType {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8> {
+        encode_i64_key_ascending(*self, id)
+    }
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self) {
+        decode_i64_key_ascending(key)
+    }
+
+    fn cmp_encoded(&self, other: &Self) -> std::cmp::Ordering {
+        self.cmp(other)
+    }
+}
+
+impl Encodable for FloatPayloadType {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8> {
+        encode_f64_key_ascending(*self, id)
+    }
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self) {
+        decode_f64_key_ascending(key)
+    }
+
+    fn cmp_encoded(&self, other: &Self) -> std::cmp::Ordering {
+        if self.is_nan() && other.is_nan() {
+            return std::cmp::Ordering::Equal;
+        }
+        if self.is_nan() {
+            return std::cmp::Ordering::Less;
+        }
+        if other.is_nan() {
+            return std::cmp::Ordering::Greater;
+        }
+        self.partial_cmp(other).unwrap()
+    }
+}
 
 pub enum NumericIndex<T: Encodable + Numericable> {
     Mutable(MutableNumericIndex<T>),
