+++ lib/segment/src/index/field_index/numeric_index/tests.rs
@@ -3,7 +3,6 @@ use rand::prelude::StdRng;
 use rand::{Rng, SeedableRng};
 use tempfile::{Builder, TempDir};
 
-use super::numeric_index_key::Encodable;
 use super::*;
 use crate::common::rocksdb_wrapper::open_db_with_existing_cf;
